backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Pydantic
pydantic==2.6.1
pydantic-settings==2.1.0

# Google APIs
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.118.0
google-genai==1.0.0

# PDF processing
pymupdf==1.23.22
pytesseract==0.3.10
pillow==10.2.0
markitdown==0.0.1a3

# Scheduler
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.1
httpx==0.26.0
